Only import docker if invoked with `dallinger docker`
